+++ chatlearn/models/vllm/hooks/vllm_0_8_5/__init__.py
@@ -15,7 +15,6 @@
 """Additional hooks of vllm-0.8.5."""
 
 from . import async_llm_engine
-# from . import input_preprocess
 from . import llm
 from . import llm_engine
 from . import loader
