+++ chatlearn/utils/vllm_utils.py
@@ -45,25 +45,14 @@ try:
 except ImportError:
     print("Cannot import megatron, please set megatron python path first.")
 
-try:
-    from chatlearn.utils.vllm_import_helper import init_world_group
-except ImportError:
-    print("Cannot import init_world_group for vLLM 0.5.1, please install vLLM 0.5.1 first.")
 
-try:
-    from chatlearn.utils.vllm_import_helper import get_pipeline_model_parallel_rank
-    from chatlearn.utils.vllm_import_helper import get_pipeline_model_parallel_world_size
-    from chatlearn.utils.vllm_import_helper import _set_default_torch_dtype
-    from chatlearn.utils.vllm_import_helper import parallel_state as mpu
-    from chatlearn.utils.vllm_import_helper import initialize_model_parallel
-    from chatlearn.utils.vllm_import_helper import initialize_dummy_weights
-except ImportError:
-    print("Cannot import vllm, please install vllm 0.3.0 or 0.5.1 first.")
+from vllm.distributed.parallel_state import init_world_group
+from vllm.model_executor.model_loader.utils import set_default_torch_dtype as _set_default_torch_dtype
+from vllm.distributed import parallel_state as mpu
+from vllm.distributed.parallel_state import initialize_model_parallel
+from vllm.model_executor.model_loader.weight_utils import initialize_dummy_weights
+from vllm.distributed.utils import get_pp_indices
 
-try:
-    from chatlearn.utils.vllm_import_helper import get_pp_indices
-except ImportError:
-    print("Cannot import vllm, please install vllm 0.6.3 first.")
 
 from .constant import QwenVersion
 
@@ -116,6 +105,17 @@ mcore_to_transformers = {
     "mlp.linear_fc2": ".mlp.down_proj.",
 }
 
+def get_model_architecture(config):
+    from vllm.model_executor.model_loader.utils import get_model_architecture  as get_model_architecture_v2
+    return get_model_architecture_v2(config)[0]
+
+
+def get_pipeline_model_parallel_rank():
+    return parallel_state.get_pp_group().rank_in_group
+
+
+def get_pipeline_model_parallel_world_size():
+    return parallel_state.get_pp_group().world_size
 
 class ParameterSyncMap:
     """Base ParameterSyncMap."""
@@ -1953,4 +1953,4 @@ def get_checkpoint_name(checkpoints_path, iteration, release=False,
     return model_path
 
 def vllm_use_v1():
-    return bool(int(os.getenv("VLLM_USE_V1", "0")))
+    return bool(int(os.getenv("VLLM_USE_V1", "1")))
